Workaround more problems with signal deferral and LLVM sanitizers
This particular glitch occurs with AddressSanitizer + LeakSanitizer.
Between SBCL's signal deferral and the sanitizer intercepting the user-
specified handlers (with different mask and flags, which seems wrong imho),
a certain test failed 5% of the time with
  "interrupt already pending when sig17 received"
And you can't run it using any out-of-process debugger, as it says:
  ==3239881==LeakSanitizer has encountered a fatal error.
  ==3239881==HINT: LeakSanitizer does not work under ptrace (strace, gdb, etc)

With this patch, which allows using the same lisp core with only C recompiled,
I got 5000 passing runs in a row. The test has to put :address-sanitizer into
*FEATURES* before save-lisp-and-die so that it will start up correctly.