Record suppressed content read just before end-of-file in sb-cover
Between the last actual top-level form in a file and the end-of-file,
there could be an arbitrary amount of *READ-SUPPRESS*ed content.  For
most uses, ignoring this is fine, but for coverage reporting where we
distinguish between uninstrumented code and code that is
conditionalized out, we need to record the suppressed content, so
return the EOF marker (which will never match any actual source path,
as it corresponds to a Top-Level Form one beyond the end of the file).

Currently untested because we need to make the internals more
accessible in order to test them.