Institute a depth cutoff for source paths
Somewhat arbitrarily choose 100, since that is where the existing test
for lp#654289 puts the dividing line between "small" and "big".  What
this means in practice is that code nested more than 100-levels deep
from the toplevel will not have accurate source path information
associated with it.

Make FORM-NUMBER-TRANSLATIONS and SUB-FIND-SOURCE-PATHS consistent in
how they register (or, rather, do not register) instances of code
while traversing the current form that are EQL to forms that have
already been seen.  This allows a consistency test on code with a
quoted constant as from the lp#654289 bug report.  Also re-enable the
compiler scaling test, since we have carefully chosen the cutoff to
make it passable.  (It's still probably a bad idea to put a literal
1000-level tree in your source code.)