Don't special case coverage-enabled code when expanding DEFMETHOD
The current implementation of the walker used in walking the method
lambda for optimizations involving slot-value (and other things)
preserves the association between original source and expanded form,
since around 2022 (commit 0c1906289ca3269806415f1bc1abceef946b9c84).
This means that we no longer have to fall back on the original method
body if storing coverage data, and can always instead use the walked
lambda.

Add a test of the coverage states in a file with a method definition
containing an optimizeable call to slot-value.