Handle read-eval more robustly in SB-COVER.
READ-EVAL can be used to inline source code using the reader.  This is
distinct from macro-based inlining (or compiler-macro, or compiler
transform...) because it happens independently of the compiler
processing the source code, and so to the source path processor there
is no distinction between the code as inserted by READ-EVAL and the
code written out long-hand.

For reporting on the state of the source code, however, there is a
significant difference: we do not want to (and indeed cannot) descend
into source code inserted by read-eval.  Define a marker struct to
denote an element read by sharp-dot, and do not attempt to descend
into it when mapping source forms to character positions.  If asked to
descend, discard that source path for state annotation
purposes (rather than returning an outer form) to avoid annotating
anything with conditional states that are irrelevant to the actual
source form.

Don't use *READ-SUPPRESS* in the replaced sharp-dot reader to avoid
having the READ-EVAL'd form annotated as having suppressed state.

Delete some unused source-mapping utilities inherited from the initial
implementation.