Fix unbounded data growth with store-coverage-data
In order to track whether a path has already emitted a coverage mark
in a given block, rather than have a hash-table keyed on paths with a
list of blocks as the value, give each block for which coverage is
emitted at all a table of source paths to act as a set.  This means
that there are no additional links to blocks during compilation, which
therefore means that they don't cause retention of their component,
which would otherwise transitively retains a whole heap of IR2
structure.

(This allows a build storing coverage marks for the system itself to
complete in a host lisp with a dynamic-space-size of 700MB, as used in
builds in github's CI.)