Avoid GENSYMing in sb-cover::make-recording-read-list
It wastes space and is ugly to look at at. While the latter criticism
is subjective, the former is not - we may wish to attach source maps
to code blobs as saved in the on-disk core. Currently the map is made
just-in-time by sb-cover, which precludes an execution model where the
instrumented code lacks read access to its own sources.
Yet, it should be feasible to dump a log file abstractly depicting a
color of each form for later rendering by an HTML frontend.